# docs/swagger.yaml
definitions:
  service.LoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  service.LoginResp:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  service.RegisterReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: Контент сервис - сервис хранения контента (баннеры, тесты, истории
    и т.д.) для дальнейшего отображения его на клиенте пользователя.
  title: Content Service API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns tokens
      parameters:
      - description: User login request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.LoginResp'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      summary: Login a user
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes the access token using the refresh token
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      summary: Refresh access token
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user in the system
      parameters:
      - description: User registration request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "409":
          description: Conflict
        "500":
          description: Internal Server Error
      summary: Register a new user
      tags:
      - Authentication
  /block:
    get:
      description: Метод получение блока контента по id и/или названию.
      parameters:
      - description: id блока
        in: query
        name: block_id
        type: integer
      - description: название блока
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получение блока контента по id и/или названию.
      tags:
      - Admin
  /block/{block_id}:
    get:
      description: Метод получения элемента контента из блока по id, названию или
        типу контента.
      parameters:
      - description: id блока
        in: path
        name: block_id
        required: true
        type: integer
      - description: id контента
        in: query
        name: content_id
        type: integer
      - description: название элемента контента
        in: query
        name: name
        type: string
      - description: тип контента (1 - баннер, 2 - истории, 3 - тесты)
        in: query
        name: content_type
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получение элементов контента из блока по id, названию или типу контента.
      tags:
      - Admin
  /block/{block_id}/list:
    get:
      description: Метод получения списка id и названий элементов контента для блока.
      parameters:
      - description: id блока
        in: path
        name: block_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получение списка id и названий элементов контента для блока.
      tags:
      - Admin
  /block/list:
    get:
      description: Метод получения списка id и названий всех блоков.
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получение списка id и названий всех блоков.
      tags:
      - Admin
...
  /template:
    get:
      description: Метод получения шаблона по id, названию или типу контента.
      parameters:
      - description: id шаблона
        in: query
        name: template_id
        type: integer
      - description: название шаблона
        in: query
        name: name
        type: string
      - description: тип шаблона (1 - баннер, 2 - истории, 3 - тесты)
        in: query
        name: content_type
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получение шаблона по id, названию или типу.
      tags:
      - Admin
  /template/list:
    get:
      description: Метод получения списка id, названий и типов всех шаблонов.
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получение списка id, названий и типов всех шаблонов.
      tags:
      - Admin
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"